chore(tooling): Initial setup and scripts
I have setup the initial setup files and scripts for the project (based on another one of my
projects) and began installing and checking necessary dependencies. This is still under review and
testing (especially with Travis-CI) and will need to go through a few iterations as well as some
build test checking.

re #2